Implemented articles lists on home page.
